flask_server/requirements.txt
flask
python-docx
flask-cors
